theme: Replace `senderName` & `recipientHeaderText` with `title`
The `senderName` & `recipientHeaderText` giving colors to name of sender
in each message and left icon + text in the Message Header of both Stream
messages and DMs, respectively, have been replaced to a lighter and darker
versions in dark and light themes, respectively, using the `title`
variable of `DesignVariables`, according to the Figma Design.

This specific UI of left icon and text in message header for the DMs
doesnt appear in the Figma design, but we use `title` variable because
`recipientHeaderText` was used previously for coloring them.

Figma:
- https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=4993-20811&t=vu2wWOBiVb0BB6Vz-0
- https://www.figma.com/design/1JTNtYo9memgW7vV6d0ygq/Zulip-Mobile?node-id=4010-6352&t=9j2mOkL1Pds0XVAE-0
discussion: https://github.com/zulip/zulip-flutter/pull/1213#discussion_r1908349307